+++ src/transformers/pipelines/__init__.py
@@ -25,8 +25,6 @@ from transformers.modelcard import ModelCard
 from transformers.models.auto.tokenization_auto import AutoTokenizer
 from transformers.pipelines.base import (
     ArgumentHandler,
-    Conversation,
-    ConversationalPipeline,
     CsvPipelineDataFormat,
     JsonPipelineDataFormat,
     PipedPipelineDataFormat,
@@ -39,6 +37,7 @@ from transformers.pipelines.base import (
     get_default_model,
     get_framework,
 )
+from transformers.pipelines.conversation import Conversation, ConversationalPipeline
 from transformers.pipelines.feature_extraction import FeatureExtractionPipeline
 from transformers.pipelines.fill_mask import FillMaskPipeline
 from transformers.pipelines.question_answering import QuestionAnsweringArgumentHandler, QuestionAnsweringPipeline
